patches.armbian/drv-phy-sun4i-usb-Allow-reset-line-to-be-shared(:1)
Original-Subject: drv:phy: sun4i-usb: Allow reset line to be shared
The USB HCIs (and PHYs?) in Allwinner's newer generation SoCs (H616)
rely on the reset line of USB PHY 2 to be de-asserted, even when only
one of the other PHYs is actually in use.

To make those ports work, we include this reset line in the HCIs' resets
property, which requires this line to be shareable.

Change the call to allocate the reset line to mark it as shared, to
enable the other ports on those SoCs.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
X-Armbian: Patch-File: patches.armbian/drv-phy-sun4i-usb-Allow-reset-line-to-be-shared
X-Armbian: Patch-File-Counter: 1
X-Armbian: Patch-Rel-Directory: patch/kernel/archive/sunxi-6.4
X-Armbian: Patch-Type: kernel
X-Armbian: Patch-Root-Type: core
X-Armbian: Patch-Sub-Type: common
X-Armbian: Original-Subject: drv:phy: sun4i-usb: Allow reset line to be shared